+++ applications/fishing-map/src/features/sidebar/heatmaps/HeatmapFilters.tsx
@@ -1,7 +1,7 @@
 import React, { Fragment, useMemo } from 'react'
-import { MultiSelect } from '@globalfishingwatch/ui-components'
 import { UrlDataviewInstance } from 'types'
 import { getFlags, getFlagsByIds } from 'utils/flags'
+import { MultiSelect } from '@globalfishingwatch/ui-components'
 import { useDataviewInstancesConnect } from 'features/workspace/workspace.hook'
 import styles from './HeatmapFilters.module.css'