+++ packages/layer-composer/src/generators/heatmap/util/time-chunks.ts
@@ -4,12 +4,22 @@ type Interval = '10days' | 'day' | 'hour'
 
 export type TimeChunk = {
   id: string
-  interval: Interval
-  intervalInDays: number
   start?: string
   viewEnd?: string
   dataEnd?: string
   quantizeOffset: number
+  frame: number
+  active: boolean
+}
+
+export type TimeChunks = {
+  chunks: TimeChunk[]
+  delta: number
+  deltaInDays: number
+  interval: Interval
+  activeStart: string
+  activeEnd: string
+  activeChunkFrame: number
 }
 
 const toDT = (dateISO: string) => DateTime.fromISO(dateISO).toUTC()
@@ -114,9 +124,10 @@ const getTimeChunks = (
   datasetStart: string,
   datasetEnd: string,
   interval: Interval,
-  intervalInDays: number
+  activeStart: string
 ) => {
   const config = CONFIG_BY_INTERVAL[interval]
+  let activeChunkFrame = 0
   const chunks: TimeChunk[] = chunkStarts.map((chunkStart) => {
     // end of *usable* tileset is end of year
     // end of *loaded* tileset is end of year + 100 days
@@ -133,18 +144,25 @@ const getTimeChunks = (
     // we will substract every timestamp with this to end up with shorter arrays indexes
     const quantizeOffset = config.getFrame(+chunkStart)
 
+    const activeStartDT = +toDT(activeStart)
+    const isActive = activeStartDT > +chunkStart && activeStartDT <= chunkViewEnd
+    const frame = toQuantizedFrame(activeStart, quantizeOffset, interval)
+    if (isActive) {
+      activeChunkFrame = frame
+    }
+
     const chunk: TimeChunk = {
-      interval,
       start,
       viewEnd,
       dataEnd,
       quantizeOffset,
       id: `heatmapchunk_${start.slice(0, 13)}_${viewEnd.slice(0, 13)}`,
-      intervalInDays,
+      frame,
+      active: isActive,
     }
     return chunk
   })
-  return chunks
+  return { chunks, activeChunkFrame }
 }
 
 /**
@@ -159,21 +177,31 @@ export const getActiveTimeChunks = (
   activeEnd: string,
   datasetStart: string,
   datasetEnd: string
-): TimeChunk[] => {
+): TimeChunks => {
   const delta = +toDT(activeEnd) - +toDT(activeStart)
-  const interval = getInterval(delta)
-  const intervalInDays = Duration.fromMillis(delta).as('days')
+  const timeChunks: TimeChunks = {
+    activeStart,
+    activeEnd,
+    chunks: [],
+    delta,
+    deltaInDays: Duration.fromMillis(delta).as('days'),
+    interval: getInterval(delta),
+    activeChunkFrame: 0,
+  }
 
   // ignore any start/end time chunk calculation as for the '10 days' interval the entire tileset is loaded
-  if (interval === '10days') {
-    return [
+  if (timeChunks.interval === '10days') {
+    const frame = toQuantizedFrame(activeStart, 0, timeChunks.interval)
+    timeChunks.chunks = [
       {
-        interval,
         quantizeOffset: 0,
         id: 'heatmapchunk_10days',
-        intervalInDays,
+        frame,
+        active: true,
       },
     ]
+    timeChunks.activeChunkFrame = frame
+    return timeChunks
   }
 
   // calculate start and end taking some buffer into account to proactively load time chunks
@@ -181,18 +209,26 @@ export const getActiveTimeChunks = (
   const bufferedActiveStart = +toDT(activeStart) - bufferSize
   const bufferedActiveEnd = +toDT(activeEnd) + bufferSize
 
-  // case when view range is completely outside of dataset range
+  // case when view range is completely outside of dataset range - no chunks
   if (bufferedActiveEnd < +toDT(datasetStart) || bufferedActiveStart > +toDT(datasetEnd)) {
-    return []
+    return timeChunks
   }
 
-  const chunkStarts = getChunkStarts(bufferedActiveStart, bufferedActiveEnd, interval)
-  const chunks = getTimeChunks(chunkStarts, datasetStart, datasetEnd, interval, intervalInDays)
+  const chunkStarts = getChunkStarts(bufferedActiveStart, bufferedActiveEnd, timeChunks.interval)
+  const { chunks, activeChunkFrame } = getTimeChunks(
+    chunkStarts,
+    datasetStart,
+    datasetEnd,
+    timeChunks.interval,
+    activeStart
+  )
+  timeChunks.chunks = chunks
+  timeChunks.activeChunkFrame = activeChunkFrame
 
-  return chunks
+  return timeChunks
 }
 
-export const toQuantizedFrame = (date: string, quantizeOffset: number, interval: Interval) => {
+const toQuantizedFrame = (date: string, quantizeOffset: number, interval: Interval) => {
   const config = CONFIG_BY_INTERVAL[interval]
   const ms = new Date(date).getTime()
   const frame = config.getFrame(ms)