+++ applications/temporalgrid-demo/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @globalfishingwatchapp/temporalgrid-demo
 
+## 0.6.10
+
+### Patch Changes
+
+- Updated dependencies [6b5fdf8]
+  - @globalfishingwatch/layer-composer@4.0.0
+  - @globalfishingwatch/react-hooks@4.0.0
+  - @globalfishingwatch/timebar@1.1.0
+
 ## 0.6.9
 
 ### Patch Changes